Move Resource constants to a new profile package
