+++ xknx/profile/__init__.py
@@ -1,0 +1,19 @@
+"""Module for Profile and resource management ."""
+# flake8: noqa
+from .const import (
+    ResourceDevicePropertyId,
+    ResourceEModeChannelPropertyId,
+    ResourceEModeDevicePropertyId,
+    ResourceGenericPropertyId,
+    ResourceGroupObjectTablePropertyId,
+    ResourceKNXNETIPPropertyId,
+    ResourceLTERoutingTablePropertyId,
+    ResourceObjectType,
+    ResourcePollingMasterInterfacePropertyId,
+    ResourceProgramPropertyId,
+    ResourcePropertyId,
+    ResourceRFMediumPropertyId,
+    ResourceRouterPropertyId,
+    ResourceSecureInterfacePropertyId,
+    ResourceTextCataloguePropertyId,
+)
